Moved preferences modal before product-search
