+++ frontend/src/pages/product-search/ProductSearch.jsx
@@ -3,7 +3,6 @@ import { useLocation, useNavigate } from "react-router-dom";
 
 import "./ProductSearch.css";
 import FeaturePriorityModal from "../../components/modals/content/feature-priority/FeaturePriorityModal";
-import PreferencesModal from "../../components/modals/content/preferences/PreferencesModal";
 import ProductOption from "../../components/product-option/ProductOption";
 import SearchBar from "../../components/search-bar/SearchBar";
 
@@ -14,9 +13,7 @@ function ProductSearch() {
     const [mainSelectedProducts, setMainSelectedProducts] = useState([]);
     const [currentSelectedProducts, setCurrentSelectedProducts] = useState([]);
 
-    const [isPreferencesModalOpen, setIsPreferencesModalOpen] = useState(false);
     const [preferencesModalData, setPreferencesModalData] = useState(null);
-
     const [isFeaturePriorityModalOpen, setIsFeaturePriorityModalOpen] = useState(false);
     const [featurePriorityModalData, setFeaturePriorityModalData] = useState(null);
 
@@ -61,11 +58,12 @@ function ProductSearch() {
 
     useEffect(() => {
         const q = location.state === null ? "" : location.state.query;
+        const preferences = location.state === null ? null : location.state.preferences;
+        setPreferencesModalData(preferences);
         _searchProductsHelper(q);
     }, [location.state]);
 
-    const onSearchSubmit = (event) => {
-        event.preventDefault();
+    const onSearchSubmit = () => {
         if (currentSelectedProducts.length !== 0) {
             setMainSelectedProducts([...mainSelectedProducts, ...currentSelectedProducts]);
             setCurrentSelectedProducts([]);
@@ -82,12 +80,6 @@ function ProductSearch() {
         } else {
             setCurrentSelectedProducts([...currentSelectedProducts, product]);
         }
-    }
-
-    const handlePreferencesModalSubmit = (data) => {
-        setPreferencesModalData(data);
-        setIsPreferencesModalOpen(false);
-        setIsFeaturePriorityModalOpen(true);
     }
 
     const handleFeaturePriorityModalSubmit = (data) => {
@@ -107,6 +99,12 @@ function ProductSearch() {
 
 	return (
 		<div>
+            <FeaturePriorityModal
+                isOpen={isFeaturePriorityModalOpen}
+                onSubmit={handleFeaturePriorityModalSubmit}
+                onClose={() => setIsFeaturePriorityModalOpen(false)}
+            />
+
             <h1>Select products to compare</h1>
             <SearchBar onSearchSubmit={onSearchSubmit} query={searchQuery} setQuery={setSearchQuery}/>
             <br/>
@@ -124,22 +122,10 @@ function ProductSearch() {
             <button>Cancel</button>
             <button
                 disabled={mainSelectedProducts.length + currentSelectedProducts.length < 3}
-                onClick={() => setIsPreferencesModalOpen(true)}
+                onClick={() => setIsFeaturePriorityModalOpen(true)}
             >Continue</button>
             <br/>
             <br/>
-
-            <PreferencesModal
-                isOpen={isPreferencesModalOpen}
-                onSubmit={handlePreferencesModalSubmit}
-                onClose={() => setIsPreferencesModalOpen(false)}
-            />
-
-            <FeaturePriorityModal
-                isOpen={isFeaturePriorityModalOpen}
-                onSubmit={handleFeaturePriorityModalSubmit}
-                onClose={() => setIsFeaturePriorityModalOpen(false)}
-            />
 
             {mainSelectedProducts.map(product => (
                 <ProductOption
